# services/platform-api/openapi/platform-api.v1.yaml
openapi: 3.1.0
info:
  title: Money Noodle Platform API
  summary: Interface-neutral Money Noodle platform contract.
  description: Canonical v1 transport contract for public and operational platform reads.
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
...
servers:
  - url: /
    description: Runtime-configured API origin.
security: []
...
paths:
...
  /v1/identity/session:
    post:
      operationId: signIn
      summary: Exchange a verified identity-provider token for a server-side session.
      description: >-
        The only sign-in path. The body carries an identity token issued by the
        configured identity provider; this API verifies it, requires that it
        proves a second factor was used for this sign-in, and establishes a
        server-side session bound to the single account.


        The response carries no token and no credential. The session is an opaque
        identifier returned in an `HttpOnly`, `Secure`, `SameSite=Strict` cookie,
        and every later request is authorised against a row this platform owns, so
        revocation is immediate and does not depend on the provider.


        A refusal is a problem document and says only which of the fixed reasons
        it was. The difference between a bad signature, a wrong audience and an
        expired token is deliberately not published.
      tags: [Identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '201':
          description: A session was established.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummary'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '503':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: getSession
      summary: Read the current session.
      description: >-
        Confirms that the caller holds an accepted session and names the account
        it is bound to. Carries no identity-provider claim and no personal data:
        the account identifier and the session's expiry are the whole of it.
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The current session.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummary'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: signOut
      summary: Revoke the current session.
      description: >-
        Sets the session's revocation time server-side and clears the cookie.
        Presenting a session that is already unknown, revoked or expired still
        succeeds, so this operation cannot be used to discover which session
        identifiers exist; presenting none at all is refused.
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        '204':
          description: The session is revoked.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/budgets:
    get:
      operationId: listBudgets
      summary: List the account's budget records.
      description: >-
        Exactly two records, `paper` and `live`, with the same schema. The
        difference between them is authority, not structure: `live` carries
        `hasExecutionAuthority: false` and nothing in this platform can act on it.
      tags: [Budgets]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The account's budget records.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetList'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '503':
          $ref: '#/components/responses/ReadModelProblem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/budgets/{kind}:
    parameters:
      - $ref: '#/components/parameters/BudgetKind'
    get:
      operationId: getBudgetDetail
      summary: Read one budget record and the control state derived from its intent.
      description: >-
        `desiredState` is what the latest `pause` or `resume` intent asked for.
        `appliedState` is what a job said it did, and is `null` until a job has
        read that intent. The two are separate fields rather than one because an
        operator's request and an engine's action are different facts, and in this
        milestone no job exists to turn the first into the second.
      tags: [Budgets]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The budget record and its derived control state.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetDetail'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/ReadModelProblem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/budgets/{kind}/controls:
    parameters:
      - $ref: '#/components/parameters/BudgetKind'
    post:
      operationId: recordBudgetControl
      summary: Record a control action against a budget as durable intent.
      description: >-
        Identical for both budgets. The action is appended to the append-only
        control table as an intent row carrying the capability, action, actor,
        epoch, recorded time and run identifier, and **this API performs no
        effect**: the job that owns the capability reads intent at the start of
        its next run.


        A 202 therefore means the request is recorded and auditable, never that
        anything has happened. For the `live` budget no job exists at all, so the
        row is recorded and audited and nothing will act on it.
      tags: [Budgets]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlRequest'
      responses:
        '202':
          description: The intent is recorded. Nothing has been performed.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/budgets/{kind}/intents:
    parameters:
      - $ref: '#/components/parameters/BudgetKind'
    get:
      operationId: getBudgetIntentHistory
      summary: Read the recorded control intent and what jobs said about it.
      description: >-
        The audit, newest first, ordered by `(epoch, recordedAt, id)` — the same
        ordering the staleness rule uses, so the first entry is the one a job
        would treat as current. Each entry carries the outcome rows a job appended
        against it, which is empty until a job has run.
      tags: [Budgets]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The bounded intent history for this budget's capability.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentHistory'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/ReadModelProblem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/engine/jobs:
    get:
      operationId: getJobHealth
      summary: Read the last recorded run of each engine job.
      description: >-
        One entry per capability the engine store knows about. A job that has
        never run reports null times and no outcome, which in this milestone is
        every job: the engine boundary is accepted and no job is built.
      tags: [Budgets]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The last recorded run of each job.
          headers:
            x-request-id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobHealthReport'
        '401':
          $ref: '#/components/responses/IdentityProblem'
        '503':
          $ref: '#/components/responses/ReadModelProblem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-mn_session
      description: >-
        An opaque server-side session identifier. It carries no claim and no
        signature: everything a request needs is in the session row this platform
        owns, so revocation is immediate and does not involve the identity
        provider. The cookie is `HttpOnly`, `Secure`, `SameSite=Strict` and
        `Path=/`, and its `__Host-` prefix makes a browser refuse it unless those
        hold.
  parameters:
    BudgetKind:
      name: kind
      in: path
      required: true
      description: Which of the account's two budget records the operation addresses.
      schema:
        $ref: '#/components/schemas/BudgetKind'
  headers:
    RequestId:
      description: Bounded diagnostic correlation identifier.
      schema:
        $ref: '#/components/schemas/RequestId'
  responses:
    IdentityProblem:
      description: >-
        The request carried no accepted session, or the session was refused. The
        `errorCode` distinguishes the reasons this API is willing to publish:
        `MN-SESSION-REQUIRED` (no session was presented), `MN-SESSION-REJECTED`
        (the session is unknown, revoked or expired), `MN-IDENTITY-REJECTED` (the
        presented identity token was not accepted),
        `MN-SECOND-FACTOR-REQUIRED` (the token did not prove a second factor was
        used) and `MN-IDENTITY-NOT-CONFIGURED` (this revision has no identity
        configuration). No provider message is ever carried.
      headers:
        x-request-id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ReadModelProblem:
      description: >-
        The read model could not answer. The `errorCode` distinguishes the three
        reasons: `MN-READ-MODEL-UNREACHABLE` (the read model could not be
        reached), `MN-READ-MODEL-NOT-PUBLISHED` (it is reachable and has
        published no such record yet), and `MN-READ-MODEL-INVALID` (it returned a
        record this API does not understand; `detail` names the field path that
        failed and no value from it).
      headers:
        x-request-id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Problem:
      description: Safe RFC 9457 problem details response.
      headers:
        x-request-id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    RequestId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
    # --- Shared read-model vocabulary ---------------------------------------
    #
    # Two rules decide how a string field is typed below, and they are worth
    # stating once rather than per field.
    #
    #   * A value the source database constrains (`venue`, `side`, `outcome`,
    #     `liquidityRole`) is a closed enum: a value outside the set is refused
    #     before it reaches a response.
    #   * A value the source database does not constrain, or that arrives inside
    #     a stored JSON document, is an open string whose known values are listed
    #     in its description. An unknown stored value must not break a client, so
    #     it is served as the string it is.
...
    OpaqueLabel:
      description: >-
        A display label chosen by the source system. Opaque to this API: it may
        contain non-ASCII punctuation and its set of values is not fixed.
      type: string
      minLength: 1
      maxLength: 200

    # --- 1. Simulated budget ------------------------------------------------
...
    SignInRequest:
      type: object
      additionalProperties: false
      required: [idToken]
      properties:
        idToken:
          description: >-
            An identity token issued by the configured provider. It is verified
            and discarded; nothing derived from it is stored beside the session.
          type: string
          minLength: 1
          maxLength: 8192
    SessionSummary:
      type: object
      additionalProperties: false
      required: [accountId, expiresAt, schemaVersion, requestId]
      properties:
        accountId:
          description: The single account this session is bound to.
          $ref: '#/components/schemas/OpaqueLabel'
        expiresAt:
          description: When this session stops being accepted, on this API's clock.
          $ref: '#/components/schemas/ApiTime'
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    BudgetList:
      type: object
      additionalProperties: false
      required: [budgets, schemaVersion, requestId]
      properties:
        budgets:
          description: Exactly two records, `paper` and `live`, in a stable order.
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/BudgetSummary'
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'
    BudgetDetail:
      type: object
      additionalProperties: false
      required:
        [
          budget,
          capability,
          desiredState,
          appliedState,
          epoch,
          latestIntentAt,
          schemaVersion,
          requestId,
        ]
      properties:
        budget:
          $ref: '#/components/schemas/BudgetSummary'
        capability:
          description: The capability intent for this budget is recorded against.
          $ref: '#/components/schemas/OpaqueLabel'
        desiredState:
          description: >-
            What the latest `pause` or `resume` intent asked for. `unset` means no
            such control has ever been recorded.
          type: string
          enum: [running, paused, unset]
        appliedState:
          description: >-
            What a job last said it did. Null until a job has read an intent row,
            which in this milestone is always.
          oneOf:
            - $ref: '#/components/schemas/IntentOutcomeState'
            - type: 'null'
        epoch:
          description: The control epoch of the latest intent row, or 0 when there is none.
          type: integer
          minimum: 0
        latestIntentAt:
          description: When the latest intent was recorded, or null when there is none.
          oneOf:
            - $ref: '#/components/schemas/ApiTime'
            - type: 'null'
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'
    ControlRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action:
          $ref: '#/components/schemas/ControlActionName'
        parameters:
          oneOf:
            - $ref: '#/components/schemas/ControlParameters'
            - type: 'null'
    ControlAccepted:
      type: object
      additionalProperties: false
      required: [intentId, capability, action, recorded, schemaVersion, requestId]
      properties:
        intentId:
          description: The identity of the appended intent row.
          $ref: '#/components/schemas/OpaqueLabel'
        capability:
          $ref: '#/components/schemas/OpaqueLabel'
        action:
          $ref: '#/components/schemas/ControlActionName'
        recorded:
          description: >-
            Always true, and always only this. The row is recorded; nothing has
            been performed, and a job will read it at the start of its next run.
          type: boolean
          const: true
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    IntentHistory:
      type: object
      additionalProperties: false
      required: [capability, entries, schemaVersion, requestId]
      properties:
        capability:
          $ref: '#/components/schemas/OpaqueLabel'
        entries:
          description: Newest first, ordered by `(epoch, recordedAt, id)`.
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/IntentHistoryEntry'
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'
...
    JobHealthReport:
      type: object
      additionalProperties: false
      required: [jobs, schemaVersion, requestId]
      properties:
        jobs:
          description: One entry per capability the engine store knows about.
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/JobHealthEntry'
        schemaVersion:
          type: string
          const: '1'
        requestId:
          $ref: '#/components/schemas/RequestId'